move test case to AbstractContentLoaderFolderJsonTest
